backend/redflags/log_detector/requirements.txt
psycopg2-binary
requests
python-dotenv
